fix some mistakes
