--- TagsCloudVisualization/README.md
@@ -1,6 +1,0 @@
-200 rectangles:
-![first example](https://github.com/5line5/tdd/blob/master/TagsCloudVisualization/examples/200.png)
-500 rectangles: 
-![second example](https://github.com/5line5/tdd/blob/master/TagsCloudVisualization/examples/500.png)
-1000 rectangles: 
-![third example](https://github.com/5line5/tdd/blob/master/TagsCloudVisualization/examples/1000.png)
