+++ TagsCloudVisualization/TagsCloudVisualization/PointExtention.cs
@@ -1,0 +1,17 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Drawing;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace TagsCloudVisualization
+{
+    public static class PointExtention
+    {
+        public static double DistanceTo(this Point p1, Point p2)
+        {
+            return Math.Sqrt(Math.Pow((p1.X - p2.X), 2) + Math.Pow((p1.Y - p2.Y), 2));
+        }
+    }
+}
