+++ TagsCloudVisualization/TagsCloudVisualization/packages.config
@@ -1,0 +1,6 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="FluentAssertions" version="5.9.0" targetFramework="net472" />
+  <package id="NUnit" version="3.12.0" targetFramework="net472" />
+  <package id="NUnit3TestAdapter" version="3.15.1" targetFramework="net472" />
+</packages>
